requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.2
SQLAlchemy>=2.0
python-dotenv>=1.0
httpx>=0.27
pytest>=8.0
pytest-cov>=5.0
